#956 Phase 4: extract cos/token_bucket.rs from tx.rs
Per docs/pr/956-phase4-token-bucket/plan.md (v7, post Codex 6 rounds
+ Gemini 2 rounds — both reviewers PLAN-READY), extract the
token-bucket lease/refill subsystem from tx.rs into cos/token_bucket.rs.

Move list (~175 LOC of code, ~220 LOC including header):

Functions:
- maybe_top_up_cos_root_lease      (#[inline], pub(in crate::afxdp))
- maybe_top_up_cos_queue_lease     (#[inline], pub(in crate::afxdp))
- refill_cos_tokens                (#[inline], pub(in crate::afxdp))
- cos_refill_ns_until              (pub(in crate::afxdp))
- release_cos_root_lease           (pub(in crate::afxdp))
- release_all_cos_root_leases      (pub(in crate::afxdp))
- release_all_cos_queue_leases     (pub(in crate::afxdp))

Constant:
- COS_MIN_BURST_BYTES              (pub(in crate::afxdp))

#[inline] is added on the 3 per-byte hot-path helpers — same pattern
Phase 2 (cos_flow_bucket_index, cos_queue_prospective_active_flows)
and Phase 3 (apply_cos_admission_ecn_policy, cos_queue_flow_share_limit,
cos_flow_aware_buffer_limit) validated. The other 4 helpers fire
once-per-drain-loop or once-per-shutdown, so they stay un-attributed.

Phase-3 forward-debt resolved: cos/admission.rs previously imported
COS_MIN_BURST_BYTES via `use crate::afxdp::tx::COS_MIN_BURST_BYTES`
(admission -> tx back-edge). After this PR admission imports via
`use super::COS_MIN_BURST_BYTES` (cos/mod.rs re-export), so both
admission.rs and tx.rs reach the constant from cos/.

worker.rs gets explicit `use super::cos::{release_all_cos_*}`
because release_all_*_leases callers all live in worker.rs and
the previous tx-glob path is gone.

tx_frame_capacity() lives in parent afxdp.rs, so token_bucket.rs
imports it via `use crate::afxdp::tx_frame_capacity;`.

Stale-text cleanup:
- tx.rs cos-imports comment block at 3472+ — now describes Phase 4
  as 'this PR' and lists the new imports
- tx.rs:3466-3472 — Phase-3 forward-debt comment for COS_MIN_BURST_BYTES
  removed (constant has moved)
- cos/mod.rs:1-7 — phase-order header bumped to current state
- cos/admission.rs:24-27 — past-tense the 'COS_MIN_BURST_BYTES stays
  in tx.rs' note

Scope vs the umbrella plan: this Phase 4 deliberately defers
cos_tick_for_ns, cos_timer_wheel_level_and_slot,
cos_guarantee_quantum_bytes, cos_surplus_quantum_bytes (which the
umbrella plan listed for Phase 4) to a future drain-scheduler phase.
Rationale documented in plan.md: those helpers depend on
COS_TIMER_WHEEL_* / COS_GUARANTEE_* constants that belong with the
park_cos_queue / advance_cos_timer_wheel scheduler entry points,
not with token-bucket lease/refill.

Net change: -175 lines from tx.rs (17198 -> 17030 lines).

Tests: cargo build --bins clean (no Phase-4 unused-import warnings);
cargo test --bins -> 865 passed, 0 failed, 2 ignored (same baseline
as Phase 3).

Reviewers signed off: Codex rounds 1-6 (PLAN-READY at v6),
Gemini rounds 1-2 (PLAN-READY at v7 after architectural #[inline]
+ scope-deferral fixes).
