+++ userspace-dp/src/afxdp/worker.rs
@@ -1,5 +1,10 @@
 use super::*;
 
+// #956 Phase 4: explicit imports for items that moved out of tx.rs into
+// cos/token_bucket.rs. Without this, the `use super::*;` glob no longer
+// reaches them (tx.rs's pub(super) re-export is gone).
+use super::cos::{release_all_cos_queue_leases, release_all_cos_root_leases};
+
 pub(crate) struct BindingWorker {
     pub(crate) slot: u32,
     pub(crate) queue_id: u32,
